Source: sneaky invariants that reach inside subnodes
Add rudimentary support for "sneaky invariants", which are invariants
that can reach inside places where normal code can't go. The idea is
that invariants often need to refer to the internal state of subnodes.
This isn't usually allowed though, because it's not the "right" way to
write programs. We relax the restriction for invariants.

When we have invariant generation techniques such as IC3, it might make
sense to write the generated invariants out as sneaky invariants, so the
user doesn't have to perform the invariant search each time.