Smt/Prove: counterexample: mark "overflows" in the trace.
This is a best-effort, and unfortunately could miss some overflows
in the internal guts of a subnode. To do this properly I think we
want a dependency analysis to find any values that depend on
overflowed values, but the whole bounded arithmetic story needs
to be re-thought a bit first.