Equivalence: start on equivalence-graph-based invariant generation
The equivalence pass is supposed to find the invariants that are
obvious on the pure, stream-based semantics, but less clear on the
transition system. The idea is that we can apply rewrites such as
```
pre (f e)   = f (pre e)
f (e -> e') = f e -> f e'
```
to saturation using an equivalence graph.

So far it only pulls out recursive bindings to a nameless form so
that `count = (0 -> pre count) + 1` becomes
`count = µ. (0 -> pre µ0) + 1` which we can compare for equality
with other recursive bindings without the names getting in the way.