Equivalence: e-graphs maintain set of canonical nodes
Ensure that the e-graphs keep their nodes map as canonical
nodes after rebuilding. I'm a bit confused about the paper
version. This isn't listed as an invariant in the paper, so
maybe it's not necessary. However, without it, you can get
non-canonical nodes when extracting the classes, and they
can never be removed from the nodes map.